Change service control method to not use "sc.exe"
In this case, we use the Monitor executable by itself, run it as elevated with some parameters to control the service and exit. This way, the UAC prompt will show GLPI Agent Monitor instead of the Windows Service Controller. Looks better. Also, added a little error displaying to the LoadStringAndMessageBox function.